Make `networkmonitor` a feature, not a service.
