+++ features/networkmonitor/impl/src/main/AndroidManifest.xml
