+++ features/roomlist/impl/build.gradle.kts
@@ -49,7 +49,7 @@ dependencies {
     implementation(projects.libraries.testtags)
     implementation(projects.libraries.uiStrings)
     implementation(projects.libraries.dateformatter.api)
-    implementation(projects.services.networkmonitor.api)
+    implementation(projects.features.networkmonitor.api)
     implementation(libs.accompanist.placeholder)
     api(projects.features.roomlist.api)
     ksp(libs.showkase.processor)
@@ -62,7 +62,7 @@ dependencies {
     testImplementation(libs.test.robolectric)
     testImplementation(projects.libraries.matrix.test)
     testImplementation(projects.libraries.dateformatter.test)
-    testImplementation(projects.services.networkmonitor.test)
+    testImplementation(projects.features.networkmonitor.test)
 
     androidTestImplementation(libs.test.junitext)
 }
