+++ plugins/src/main/kotlin/extension/DependencyHandleScope.kt
@@ -84,7 +84,6 @@ fun DependencyHandlerScope.allLibrariesImpl() {
 fun DependencyHandlerScope.allServicesImpl() {
     implementation(project(":services:analytics:noop"))
     implementation(project(":services:appnavstate:impl"))
-    implementation(project(":services:networkmonitor:impl"))
     implementation(project(":services:toolbox:impl"))
 }
