--- services/networkmonitor/test/src/main/kotlin/io/element/android/services/networkmonitor/test/FakeNetworkMonitor.kt
@@ -1,24 +1,0 @@
-/*
- * Copyright (c) 2023 New Vector Ltd
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package io.element.android.services.networkmonitor.test
-
-import io.element.android.services.networkmonitor.api.NetworkMonitor
-import io.element.android.services.networkmonitor.api.NetworkStatus
-import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.emptyFlow
-
-class FakeNetworkMonitor(override val connectivity: Flow<NetworkStatus> = emptyFlow()) : NetworkMonitor
